Wait for ViaVersion inside of Bevy
Instead of waiting for ViaVersion to return the "ready" message and continuing, create the task and poll it later. The plugin trait already has a method to determine if the plugin is "ready", so there the channel is polled for the result of the proxy task. If an error is returned, exit the app.

Use a `oneshot` channel instead of a `watch` channel.

Moved bare functions into "helper" structs to clean things up.